backend/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
python-json-logger==2.0.7
uvicorn==0.30.0
prospector==1.10.3
pre-commit==3.7.1
Sphinx==7.3.7
renku-sphinx-theme==0.4.0
